Can transfer from guest cart to customer cart at login

+++ app/helpers/application_helper.rb
@@ -39,13 +39,16 @@ module ApplicationHelper
   end
 
   def transfer_guest_cart_to_customer_cart
-    if session[:cart]
-      guest_cart = Cart.find(session[:cart])
-      guest_cart.cart_contents.each { |content| CartContent.create(cart: current_shopping_cart.id, item: cart_content.item )}
-      CartContent.where(cart: guest_cart.id).delete_all
-      guest_cart.destroy
-      session[:shopping_cart] = nil
+    if signed_in?
+      if session[:cart]
+        guest_cart = Cart.find(session[:cart])
+        guest_cart.cart_contents.each do |content|
+          CartContent.create(cart_id: current_shopping_cart.id, item_id: content.item.id, item_quantity: content.item_quantity )
+        end
+        CartContent.where(cart_id: guest_cart.id).delete_all
+        guest_cart.destroy
+        session[:cart] = nil
+      end
     end
   end
-
 end
